OO-3296: start implementation administrator and other roles per organization, implement segment view for user visit card to better inline in user management, implements several features for all roles (allow enrollment, mandatory confirmation email by enrollment, allow autocomplete, bulk changes...), add roles like lecture manager, quality manager..., enhance the roles editor in user management

+++ src/main/java/org/olat/admin/user/bulkChange/UserBulkChangeManager.java
@@ -214,7 +214,7 @@ public class UserBulkChangeManager implements InitializingBean {
 
 			// set status
 			if (userBulkChanges.getStatus() != null) {
-				Integer status = userBulkChanges.getStatus().intValue();
+				Integer status = userBulkChanges.getStatus();
 
 				int oldStatus = identity.getStatus();
 				String oldStatusText = (oldStatus == Identity.STATUS_PERMANENT ? "permanent" : (oldStatus == Identity.STATUS_ACTIV ? "active" : (oldStatus == Identity.STATUS_LOGIN_DENIED ? "login_denied" : (oldStatus == Identity.STATUS_DELETED ? "deleted" : "unknown"))));
